FINAL WORKING GRAPHCONTAINER YAYAYAYAY
